+++ src/commands/profile.ts
@@ -1,5 +1,4 @@
 import type { User } from 'discord.js';
-import * as logger from '../logger';
 import { DiscordAPIError, EmbedBuilder, SlashCommandBuilder, userMention } from 'discord.js';
 import { DiscordErrorCode } from '../helpers/DiscordErrorCode';
 
@@ -15,14 +14,14 @@ const builder = new SlashCommandBuilder()
 export const profile: GlobalCommand = {
 	info: builder,
 	requiresGuild: false,
-	async execute({ client, user, interaction, reply, replyPrivately, guild, source }) {
+	async execute({ client, user, interaction, reply, guild, source }) {
 		const otherUser = interaction.options.getUser(UserParamName, false);
 
 		if (otherUser) {
 			if (source === 'dm') {
 				// Shouldn't try to snoop other users privately. Only bot or self.
 				if (otherUser.id !== user.id && otherUser.id !== client.user.id) {
-					return await replyPrivately("That user isn't here!");
+					throw new Error("That user isn't here!");
 				}
 			} else {
 				// Shouldn't try to snoop users that aren't in the guild
@@ -31,18 +30,14 @@ export const profile: GlobalCommand = {
 				} catch (error) {
 					// Throw unrelated errors
 					if (!(error instanceof DiscordAPIError)) {
-						// I know we're improving our error handling. Just throw this error once we have that in place:
-						logger.error(error);
-						return await replyPrivately('Something went wrong.');
+						throw error;
 					}
 					if (error.code !== DiscordErrorCode.UNKNOWN_MEMBER) {
-						// I know we're improving our error handling. Just throw this error once we have that in place:
-						logger.error(error);
-						return await replyPrivately('An unhandled API error happened.');
+						throw error;
 					}
 
 					// If error says that the user isn't a guild member, complain.
-					return await replyPrivately("That user isn't here!");
+					throw new Error("That user isn't here!");
 				}
 			}
 		}
@@ -51,7 +46,7 @@ export const profile: GlobalCommand = {
 
 		const url = target.avatarURL({ extension: 'png', size: 2048 });
 		if (url === null || !url) {
-			return await replyPrivately("This user doesn't seem to have an avatar!");
+			throw new Error("This user doesn't seem to have an avatar!");
 		}
 
 		// Ping the target user if not self