+++ CHANGELOG.md
@@ -5,9 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## 0.6.0 - 2022-12-03
+### Added
+- A new `/talk` command has been added - uses the [dectalk](https://github.com/babakinha/dectalk) TTS engine to speak a given message.
+- A new `Talk` context menu command has been added - also uses dectalk to speak any message users right-click on.
+
+### Changed
+- The error reporting system for commands has also been standardized.
+
+## 0.5.2 - 2022-12-02
+### Changed
+- New safety guards for the `/profile` command.
+
 ## 0.5.1 - 2022-11-29
 ### Changed
 - Use the new-and-improved [FixTweet](https://github.com/FixTweet/FixTweet) instead of ye olde [BetterTwitFix](https://github.com/dylanpdx/BetterTwitFix).
+- Our development environment is now standardized with Docker.
 
 ## 0.5.0 - 2022-10-13
 ### Added